move PlaceInfo outside Swipe
