+++ src/client/components/PlaceInfo/PlaceInfo.jsx
@@ -3,8 +3,8 @@ import { Button, Modal } from 'react-bootstrap';
 import _ from 'lodash';
 import axios from 'axios';
 
-import APIS from '../../../constants/apis';
-import PATHS from '../../../constants/paths';
+import APIS from '../../constants/apis';
+import PATHS from '../../constants/paths';
 
 const SAMPLE_PLACE = {
   place_id: 203,
