+++ src/client/components/PlaceInfo/index.js
